feat: update farm access

+++ src/core/use-cases/delete-farm-image.ts
@@ -1,5 +1,6 @@
 // Repositories
 import { FarmsRepository } from "@/core/repositories/farms-repository";
+import { UsersRepository } from "@/core/repositories/users-repository";
 
 // Services
 import { Storage } from "@/core/storage/storage";
@@ -8,6 +9,7 @@ import { Storage } from "@/core/storage/storage";
 import { ResourceNotFoundError } from "@/core/errors/resource-not-found";
 
 interface DeleteFarmImageUseCaseRequest {
+  user_id: string;
   farm_id: string;
   image_url: string;
 }
@@ -15,10 +17,15 @@ interface DeleteFarmImageUseCaseRequest {
 export class DeleteFarmImageUseCase {
   constructor(
     private farmsRepository: FarmsRepository,
+    private usersRepository: UsersRepository,
     private storage: Storage
   ) {}
 
-  async execute({ farm_id, image_url }: DeleteFarmImageUseCaseRequest) {
+  async execute({
+    user_id,
+    farm_id,
+    image_url,
+  }: DeleteFarmImageUseCaseRequest) {
     const farm = await this.farmsRepository.find("basic", { id: farm_id });
 
     if (!farm) throw new ResourceNotFoundError("Fazenda", farm_id);
@@ -26,6 +33,13 @@ export class DeleteFarmImageUseCase {
     const image = farm.images.get(image_url);
 
     if (!image) throw new ResourceNotFoundError("Imagem", image_url);
+
+    const user = await this.usersRepository.find("basic", { id: user_id });
+
+    if (!user) throw new ResourceNotFoundError("Usuário", user_id);
+
+    if (!farm.admin_id.equals(user.id) && !user.admin)
+      throw new ResourceNotFoundError("Fazenda", farm_id);
 
     await this.storage.delete(image, "farms");
